# pkg/swagger/docs/swagger.yaml
definitions:
...
  emplacc-api_api_v1_dto.ErrorDetail:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  emplacc-api_api_v1_dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorDetail'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
  emplacc-api_api_v1_dto.ForumMessageResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.ForumMessage'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
  emplacc-api_api_v1_dto.ForumMessagesListResponse:
    properties:
      data:
        $ref: '#/definitions/emplacc-api_api_v1_dto_response.ForumMessagesPage'
      meta:
        $ref: '#/definitions/emplacc-api_api_v1_dto.ResponseMeta'
    type: object
...
  emplacc-api_api_v1_dto.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
...
  emplacc-api_api_v1_dto.ResponseMeta:
    properties:
      pagination:
        $ref: '#/definitions/emplacc-api_api_v1_dto.Pagination'
      trace_id:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_request.CreateForumMessage:
    properties:
      creator_id:
        type: string
      description:
        items:
          type: string
        type: array
      problem_id:
        type: string
    type: object
...
  emplacc-api_api_v1_dto_request.UpdateForumMessage:
    properties:
      description:
        items:
          type: string
        type: array
    type: object
...
  emplacc-api_api_v1_dto_response.ForumMessage:
    properties:
      created_at:
        type: string
      creator_id:
        type: string
      description:
        items:
          type: string
        type: array
      id:
        type: string
      problem_id:
        type: string
      updated_at:
        type: string
    type: object
  emplacc-api_api_v1_dto_response.ForumMessagesPage:
    properties:
      messages:
        items:
          $ref: '#/definitions/emplacc-api_api_v1_dto_response.ForumMessage'
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /forum-messages:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of forum messages
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ForumMessagesListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Forum Messages
      tags:
      - ForumMessages
    post:
      consumes:
      - application/json
      description: Create a new forum message
      parameters:
      - description: Forum Message data
        in: body
        name: forum_message
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.CreateForumMessage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ForumMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Create Forum Message
      tags:
      - ForumMessages
  /forum-messages/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a forum message by its ID
      parameters:
      - description: Forum Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Delete Forum Message
      tags:
      - ForumMessages
    get:
      consumes:
      - application/json
      description: Retrieve a specific forum message by its ID
      parameters:
      - description: Forum Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ForumMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Get Forum Message
      tags:
      - ForumMessages
    patch:
      consumes:
      - application/json
      description: Update an existing forum message
      parameters:
      - description: Forum Message ID
        in: path
        name: id
        required: true
        type: string
      - description: Forum Message data
        in: body
        name: forum_message
        required: true
        schema:
          $ref: '#/definitions/emplacc-api_api_v1_dto_request.UpdateForumMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ForumMessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: Update Forum Message
      tags:
      - ForumMessages
...
  /problems/{problem_id}/forum-messages:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of forum messages for a specific problem
      parameters:
      - description: Problem ID
        in: path
        name: problem_id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ForumMessagesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/emplacc-api_api_v1_dto.ErrorResponse'
      summary: List Forum Messages by Problem
      tags:
      - ForumMessages
...
swagger: "2.0"